+++ dila/data/__init__.py
@@ -1,4 +1,4 @@
-from dila.data.base_strings import add_translated_string
+from dila.data.base_strings import add_or_update_base_string
 from dila.data.translated_strings import get_translated_string
 from dila.data.translated_strings import get_translated_strings
 from dila.data.translated_strings import set_translated_string
@@ -13,6 +13,6 @@ from dila.data.engine import shutdown_session
 
 
 __all__ = [
-    add_translated_string, get_translated_string, get_translated_strings, set_translated_string, add_resource,
+    add_or_update_base_string, get_translated_string, get_translated_strings, set_translated_string, add_resource,
     get_resource, get_resources, setup, shutdown_session, add_language, get_languages, get_language,
 ]